feat: Autenticação rotas de LoanController e validação da rota de criar Loan

+++ Controllers/LoansController.cs
@@ -4,21 +4,26 @@ using ApiBiblioteca.Application.Utils;
 using ApiBiblioteca.Application.ViewModel;
 using AutoMapper;
 using ApiBiblioteca.Domain.DTOs;
-using ApiBiblioteca.Domain.Models;
+using Microsoft.AspNetCore.Authorization;
+using ApiBiblioteca.Application.Services;
+using ApiBiblioteca.Infra.Repositories;
 
 namespace ApiBiblioteca.Controllers
 {
+    [Authorize]
     [Route("api/v1/loans")]
     [ApiController]
     public class LoansController : ControllerBase
     {
         private readonly ILoanRepository _loanRepository;
         private readonly IMapper _mapper;
+        private readonly TokenService _tokenService;
 
-        public LoansController(ILoanRepository loanRepository, IMapper mapper)
+        public LoansController(ILoanRepository loanRepository, IMapper mapper, TokenService tokenService)
         {
             _loanRepository = loanRepository;
             _mapper = mapper;
+            _tokenService = tokenService;
         }
 
         [HttpGet]
@@ -42,8 +47,11 @@ namespace ApiBiblioteca.Controllers
         [HttpPost]
         public async Task<IActionResult> Add([FromBody] LoanViewModel viewModel)
         {
+            var userId = _tokenService.GetIdByToken(HttpContext);
+
             var loan = CreateUtil.LoanCreate(viewModel);
-            loan.UserId = Guid.Parse("4DFC2E71-4C5B-490F-509C-08DC3C4E200D");
+            loan.UserId = Guid.Parse(userId);
+            
             await _loanRepository.Add(loan);
 
             return StatusCode(201, "Loan registered successfully!");
@@ -52,6 +60,9 @@ namespace ApiBiblioteca.Controllers
         [HttpPut("{id}")]
         public async Task<IActionResult> Update([FromBody] LoanViewModel viewModel, Guid id)
         {
+            var loanExists = await _loanRepository.GetById(id);
+            if (loanExists == null) return NotFound("Loan not found!");
+
             var loan = CreateUtil.LoanCreate(viewModel);
             await _loanRepository.Update(id, loan);
 
@@ -61,6 +72,9 @@ namespace ApiBiblioteca.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            var loanExists = await _loanRepository.GetById(id);
+            if (loanExists == null) return NotFound("Loan not found!");
+
             await _loanRepository.Delete(id);
 
             return Ok("Loan removed successfully!");
